requirements.txt
Flask==3.1.1
openai==1.90.0
python-dotenv==1.1.0
Requests==2.32.4
scikit_learn==1.7.0
Unidecode==1.3.8
